requirements.txt
python-telegram-bot==21.6
aiofiles==24.1.0
python-dotenv==1.0.1